Set up sequelizeCLI and begin to set up models
