+++ .gitignore
@@ -46,3 +46,4 @@ Thumbs.db
 
 # Environment Variables
 /server/.env
+/server/server/config/config.json
