+++ package.json
@@ -59,10 +59,11 @@
     "ngx-rating": "0.0.9",
     "normalize.css": "^7.0.0",
     "pg": "^7.4.1",
+    "pg-hstore": "^2.3.2",
     "primeng": "^5.2.0",
     "roboto-fontface": "^0.8.0",
     "rxjs": "^5.5.6",
-    "sequelize": "^4.33.1",
+    "sequelize": "^4.33.4",
     "zone.js": "^0.8.20"
   },
   "devDependencies": {
