+++ server/.sequelizerc
@@ -1,8 +1,8 @@
-var path = require('path');
+const path = require('path');
 
 module.exports = {
-  'config': path.resolve('./config', 'config.json'),
-  'migrations-path': path.resolve('./config', 'migrations'),
-  'models-path': path.resolve('./features', 'models'),
-  'seeders-path': path.resolve('./config', 'seeders')
-}
+  "config": path.resolve('./config', 'config.json'),
+  "models-path": path.resolve('./models'),
+  "seeders-path": path.resolve('./seeders'),
+  "migrations-path": path.resolve('./migrations')
+};
