+++ server/bin/www
@@ -1,0 +1,9 @@
+const http = require('http');
+const app = require('../server.js');
+
+
+const port = parseInt(process.env.PORT, 10) || 3000;
+app.set('port', port);
+
+const server = http.createServer(app);
+server.listen(port, () => {console.log(`Now listening on port ${port}`)});
