--- server/config/config.js
@@ -1,15 +1,0 @@
-development: {
-  username = secrets.dev.username,
-  password = secrets.dev.password,
-  database = secrets.dev.database,
-  host = secrets.dev.host,
-  dialect = secrets.dev.dialect
-};
-
-fi_01: {
-  username = secrets.fi_01.username,
-  password = secrets.fi_01.password,
-  database = secrets.fi_01.database,
-  host = secrets.fi_01.host,
-  dialect = secrets.fi_01.dialect
-};
