+++ server/controllers/index.js
@@ -1,0 +1,9 @@
+const authController = require('./authController');
+const userController = require('./userController');
+const feedbackController = require('./feedbackController');
+
+module.exports = {
+  authController,
+  userController,
+  feedbackController
+};
