+++ server/controllers/userController.js
@@ -1,4 +1,4 @@
-const User = require('../../models/User');
+const User = require('../models/index').user;
 
 module.exports = {
