--- server/features/routes/auth/authRoutes.js
@@ -1,9 +1,0 @@
-const authController = require('./authController');
-
-module.exports = (app) => {
-
-    app.post('/userAuth', authController.login);
-
-    app.post('/createUser', authController.createUser);
-
-};
