--- server/features/routes/user/userRoutes.js
@@ -1,11 +1,0 @@
-const userController = require('./userController'),
-      authController = require('../auth/authController');
-
-module.exports = (app) => {
-
-    app.put('/updateUser', authController.verifyValidToken, userController.updateUser);
-
-    app.get('/getUser', authController.verifyValidToken, userController.getUser);
-
-
-};
