+++ server/routes/authRoutes.js
@@ -1,0 +1,8 @@
+const authController = require('../controllers/authController');
+
+module.exports = (app) => {
+
+  app.post('/userAuth', authController.login);
+
+  app.post('/createUser', authController.createUser);
+};
