+++ server/routes/userRoutes.js
@@ -1,0 +1,8 @@
+const authController = require('../controllers/authController');
+const userController = require('../controllers/userController');
+
+module.exports = (app) => {
+  app.put('/updateUser', authController.verifyValidToken, userController.updateUser);
+
+  app.get('/getUser', authController.verifyValidToken, userController.getUser);
+};
