+++ server/server.js
@@ -1,10 +1,8 @@
 const express = require('express'),
       app = express(),
-      http = require('http'),
-      port = process.env.PORT || '3000',
-      path = require('path');
-const bodyParser = require('body-parser');
-const cookieParser = require('cookie-parser');
+      path = require('path'),
+      bodyParser = require('body-parser'),
+      cookieParser = require('cookie-parser');
 
 //Express Middleware
 app.use(bodyParser.json());
@@ -12,8 +10,9 @@ app.use(bodyParser.urlencoded({ extended: true}));
 app.use(cookieParser());
 
 //Routes
-require('./features/routes/auth/authRoutes')(app);
-require('./features/routes/user/userRoutes')(app);
+require('./routes/authRoutes')(app);
+require('./routes/userRoutes')(app);
+require('./routes/feedbackRoutes')(app);
 
 // Angular DIST output folder
 app.use(express.static(path.join(__dirname, '/../dist')));
@@ -23,10 +22,7 @@ app.get('*', (req, res) => {
   res.sendFile(path.join(__dirname, '/../dist/index.html'));
 });
 
-//Create Server
-app.set('port', port);
-const server = http.createServer(app);
-server.listen(port, () => console.log(`Running on localhost:${port}`));
+module.exports = app;
